move visibility corner case tests to server visiblity unit tests
